- Renaming the class file to 'CustomQueueNames' - Added an example for literal queueu name in queueNameExpression property - Removed the ack code to keep it simple
