+++ cloud-stream-dynamic-destination-processor/src/main/resources/application.yml
@@ -30,7 +30,7 @@ spring:
           environment:
             solace:
               java:
-                host: tcp://localhost:55555
+                host: tcp://localhost:55554
                 msgVpn: default
                 clientUsername: default
                 clientPassword: default
